feat(i18n): extract plural rules + pin Arabic mapping (issue #389)
